refactor(branches): extract HTML generation logic to BranchHtmlGenerator
Move the HTML generation logic for branches and error pages into a separate BranchHtmlGenerator class to improve code maintainability and separation of concerns. This reduces the complexity of the BranchViewProvider class and makes the code easier to manage.
